record.receivers: refactor 'populate' methods
Moved them to utils, leaving only actual receivers in the receivers
module.
