+++ tests/unit/records/test_records_receivers.py
@@ -30,337 +30,7 @@ from inspire_schemas.api import load_schema, validate
 from inspirehep.modules.records.receivers import (
     assign_phonetic_block,
     assign_uuid,
-    populate_abstract_source_suggest,
-    populate_affiliation_suggest,
-    populate_authors_name_variations,
-    populate_bookautocomplete,
-    populate_experiment_suggest,
-    populate_inspire_document_type,
-    populate_recid_from_ref,
-    populate_title_suggest,
-    populate_author_count,
-    populate_authors_full_name_unicode_normalized,
 )
-from inspire_utils.name import generate_name_variations
-
-
-def test_populate_authors_name_variations():
-    schema = load_schema('authors')
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/authors.json',
-        'name': {'value': 'Silk, James Brian'},
-        '_collections': ['Authors'],
-    }
-    assert validate(record, schema) is None
-
-    populate_authors_name_variations(None, record)
-
-    expected = generate_name_variations(record['name'].get('value'))
-    result = record['name_variations']
-
-    assert expected == result
-
-
-def test_populate_authors_name_variations_does_nothing_if_other_schema():
-    schema = load_schema('hep')
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        '_collections': ['Literature'],
-        'authors': [
-            {'full_name': 'Rafelski, Johann'},
-        ],
-        'document_type': [
-            'book',
-        ],
-        'titles': [
-            {'title': 'Relativity Matters'},
-        ],
-    }
-    assert validate(record, schema) is None
-
-    populate_authors_name_variations(None, record)
-
-    assert 'name_variations' not in record
-
-
-def test_populate_bookautocomplete_from_authors():
-    schema = load_schema('hep')
-    authors_schema = schema['properties']['authors']
-    document_type_schema = schema['properties']['document_type']
-    self_schema = schema['properties']['self']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'authors': [
-            {'full_name': 'Rafelski, Johann'},
-        ],
-        'document_type': [
-            'book',
-        ],
-        'self': {
-            '$ref': 'http://localhost:5000/api/literature/1519486',
-        },
-    }
-    assert validate(record['authors'], authors_schema) is None
-    assert validate(record['document_type'], document_type_schema) is None
-    assert validate(record['self'], self_schema) is None
-
-    populate_bookautocomplete(None, record)
-
-    expected = {
-        'input': [
-            'Rafelski, Johann',
-        ],
-    }
-    result = record['bookautocomplete']
-
-    assert expected == result
-
-
-def test_populate_bookautocomplete_from_titles():
-    schema = load_schema('hep')
-    document_type_schema = schema['properties']['document_type']
-    self_schema = schema['properties']['self']
-    titles_schema = schema['properties']['titles']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'document_type': [
-            'book',
-        ],
-        'self': {
-            '$ref': 'http://localhost:5000/api/literature/1519486',
-        },
-        'titles': [
-            {'title': 'Relativity Matters'},
-        ],
-    }
-    assert validate(record['document_type'], document_type_schema) is None
-    assert validate(record['self'], self_schema) is None
-    assert validate(record['titles'], titles_schema) is None
-
-    populate_bookautocomplete(None, record)
-
-    expected = {
-        'input': [
-            'Relativity Matters',
-        ],
-    }
-    result = record['bookautocomplete']
-
-    assert expected == result
-
-
-def test_populate_bookautocomplete_from_imprints_dates():
-    schema = load_schema('hep')
-    document_type_schema = schema['properties']['document_type']
-    self_schema = schema['properties']['self']
-    imprints_schema = schema['properties']['imprints']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'document_type': [
-            'book',
-        ],
-        'imprints': [
-            {'date': '2010-07-23'},
-        ],
-        'self': {
-            '$ref': 'http://localhost:5000/api/literature/1519486',
-        },
-    }
-    assert validate(record['document_type'], document_type_schema) is None
-    assert validate(record['imprints'], imprints_schema) is None
-    assert validate(record['self'], self_schema) is None
-
-    populate_bookautocomplete(None, record)
-
-    expected = {
-        'input': [
-            '2010-07-23',
-        ],
-    }
-    result = record['bookautocomplete']
-
-    assert expected == result
-
-
-def test_populate_bookautocomplete_from_imprints_publishers():
-    schema = load_schema('hep')
-    document_type_schema = schema['properties']['document_type']
-    self_schema = schema['properties']['self']
-    imprints_schema = schema['properties']['imprints']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'document_type': [
-            'book',
-        ],
-        'imprints': [
-            {'publisher': 'Springer'},
-        ],
-        'self': {
-            '$ref': 'http://localhost:5000/api/literature/1519486',
-        },
-    }
-    assert validate(record['document_type'], document_type_schema) is None
-    assert validate(record['imprints'], imprints_schema) is None
-    assert validate(record['self'], self_schema) is None
-
-    populate_bookautocomplete(None, record)
-
-    expected = {
-        'input': [
-            'Springer',
-        ]
-    }
-    result = record['bookautocomplete']
-
-    assert expected == result
-
-
-def test_populate_bookautocomplete_from_isbns_values():
-    schema = load_schema('hep')
-    document_type_schema = schema['properties']['document_type']
-    self_schema = schema['properties']['self']
-    isbns_schema = schema['properties']['isbns']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'document_type': [
-            'book',
-        ],
-        'isbns': [
-            {'value': '0201021153'},
-        ],
-        'self': {
-            '$ref': 'http://localhost:5000/api/literature/1519486',
-        },
-    }
-    assert validate(record['document_type'], document_type_schema) is None
-    assert validate(record['isbns'], isbns_schema) is None
-    assert validate(record['self'], self_schema) is None
-
-    populate_bookautocomplete(None, record)
-
-    expected = {
-        'input': [
-            '0201021153',
-        ],
-    }
-    result = record['bookautocomplete']
-
-    assert expected == result
-
-
-def test_populate_bookautocomplete_does_nothing_if_record_is_not_literature():
-    record = {'$schema': 'http://localhost:5000/schemas/records/other.json'}
-
-    populate_bookautocomplete(None, record)
-
-    assert 'bookautocomplete' not in record
-
-
-def test_populate_bookautocomplete_does_nothing_if_record_is_not_a_book():
-    schema = load_schema('hep')
-    authors_schema = schema['properties']['authors']
-    document_type_schema = schema['properties']['document_type']
-    self_schema = schema['properties']['self']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'authors': [
-            {'full_name': 'Mohayai, Tanaz Angelina'},
-        ],
-        'document_type': [
-            'article',
-        ],
-        'self': {
-            '$ref': 'http://localhost:5000/api/literature/1520027',
-        }
-    }
-    assert validate(record['authors'], authors_schema) is None
-    assert validate(record['document_type'], document_type_schema) is None
-    assert validate(record['self'], self_schema) is None
-
-    populate_bookautocomplete(None, record)
-
-    assert 'bookautocomplete' not in record
-
-
-def test_populate_experiment_suggest():
-    schema = load_schema('experiments')
-    legacy_name_schema = schema['properties']['legacy_name']
-    long_name_schema = schema['properties']['long_name']
-    name_variants_schema = schema['properties']['name_variants']
-    collaboration_schema = schema['properties']['collaboration']
-    accelerator_schema = schema['properties']['accelerator']
-    experiment_schema = schema['properties']['experiment']
-    institutions_schema = schema['properties']['institutions']
-
-    record = {
-        '$schema': 'http://foo/experiments.json',
-        'self': {'$ref': 'https://localhost:5000/api/experiments/bar'},
-        'legacy_name': 'foo',
-        'long_name': 'foobarbaz',
-        'name_variants': [
-            'bar',
-            'baz',
-        ],
-        'collaboration': {
-            'value': 'D0',
-        },
-        'accelerator': {
-            'value': 'LHC',
-        },
-        'experiment': {
-            'short_name': 'SHINE',
-            'value': 'NA61',
-        },
-        'institutions': [
-            {
-                'value': 'ICN',
-            },
-        ],
-    }
-
-    assert validate(record['legacy_name'], legacy_name_schema) is None
-    assert validate(record['long_name'], long_name_schema) is None
-    assert validate(record['name_variants'], name_variants_schema) is None
-    assert validate(record['collaboration'], collaboration_schema) is None
-    assert validate(record['accelerator'], accelerator_schema) is None
-    assert validate(record['institutions'], institutions_schema) is None
-    assert validate(record['experiment'], experiment_schema) is None
-
-    populate_experiment_suggest(None, record)
-
-    expected = {
-        'input': [
-            'LHC',
-            'D0',
-            'SHINE',
-            'NA61',
-            'ICN',
-            'foo',
-            'foobarbaz',
-            'bar',
-            'baz',
-        ]
-    }
-
-    result = record['experiment_suggest']
-
-    assert expected == result
-
-
-def test_populate_experiment_suggest_does_nothing_if_record_is_not_experiment():
-    record = {'$schema': 'http://localhost:5000/schemas/records/other.json'}
-
-    populate_experiment_suggest(None, record)
-
-    assert 'experiment_suggest' not in record
 
 
 def test_assign_phonetic_block_handles_ascii_names():
@@ -584,506 +254,4 @@ def test_assign_uuid_does_not_touch_existing_uuids(mock_uuid4):
     result = record['authors']
 
     assert validate(result, subschema) is None
-    assert expected == result
-
-
-def test_populate_inspire_document_type_from_document_type():
-    schema = load_schema('hep')
-    subschema = schema['properties']['document_type']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'document_type': ['thesis'],
-    }
-    assert validate(record['document_type'], subschema) is None
-
-    populate_inspire_document_type(None, record)
-
-    expected = [
-        'thesis',
-    ]
-    result = record['facet_inspire_doc_type']
-
-    assert expected == result
-
-
-def test_populate_inspire_document_type_from_refereed():
-    schema = load_schema('hep')
-    document_type_schema = schema['properties']['document_type']
-    refereed_schema = schema['properties']['refereed']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'document_type': [
-            'article',
-        ],
-        'refereed': True,
-    }
-    assert validate(record['document_type'], document_type_schema) is None
-    assert validate(record['refereed'], refereed_schema) is None
-
-    populate_inspire_document_type(None, record)
-
-    expected = [
-        'article',
-        'peer reviewed',
-    ]
-    result = record['facet_inspire_doc_type']
-
-    assert expected == result
-
-
-def test_populate_inspire_document_type_from_publication_type():
-    schema = load_schema('hep')
-    document_type_schema = schema['properties']['document_type']
-    publication_type_schema = schema['properties']['publication_type']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'document_type': [
-            'article',
-        ],
-        'publication_type': [
-            'introductory',
-        ],
-    }
-    assert validate(record['document_type'], document_type_schema) is None
-    assert validate(record['publication_type'], publication_type_schema) is None
-
-    populate_inspire_document_type(None, record)
-
-    expected = [
-        'article',
-        'introductory',
-    ]
-    result = record['facet_inspire_doc_type']
-
-    assert expected == result
-
-
-def test_populate_inspire_document_type_does_nothing_if_record_is_not_literature():
-    record = {'$schema': 'http://localhost:5000/schemas/records/other.json'}
-
-    populate_inspire_document_type(None, record)
-
-    assert 'facet_inspire_doc_type' not in record
-
-
-def test_populate_recid_from_ref():
-    json_dict = {
-        'simple_key': {'$ref': 'http://x/y/1'},
-        'key_with_record': {'$ref': 'http://x/y/2'},
-        'record': {'$ref': 'http://x/y/3'},
-        'embedded_list': [{'record': {'$ref': 'http://x/y/4'}}],
-        'embedded_record': {'record': {'$ref': 'http://x/y/5'}}
-    }
-
-    populate_recid_from_ref(None, json_dict)
-
-    assert json_dict['simple_key_recid'] == 1
-    assert json_dict['key_with_recid'] == 2
-    assert json_dict['recid'] == 3
-    assert json_dict['embedded_list'][0]['recid'] == 4
-    assert json_dict['embedded_record']['recid'] == 5
-
-
-def test_populate_recid_from_ref_handles_deleted_records():
-    json_dict = {
-        'deleted_records': [
-            {'$ref': 'http://x/y/1'},
-            {'$ref': 'http://x/y/2'},
-        ],
-    }
-
-    populate_recid_from_ref(None, json_dict)
-
-    assert json_dict['deleted_recids'] == [1, 2]
-
-
-def test_populate_abstract_source_suggest():
-    schema = load_schema('hep')
-    subschema = schema['properties']['abstracts']
-
-    record = {
-        '$schema': 'http://localhost:5000/schemas/records/hep.json',
-        'abstracts': [
-            {
-                'source': 'foo',
-                'value': 'bar',
-            },
-        ],
-    }
-    assert validate(record['abstracts'], subschema) is None
-
-    populate_abstract_source_suggest(None, record)
-
-    expected = [
-        {
-            'abstract_source_suggest': {
-                'input': 'foo',
-            },
-            'source': 'foo',
-            'value': 'bar',
-        },
-    ]
-    result = record['abstracts']
-
-    assert expected == result
-
-
-def test_populate_abstract_source_suggest_does_nothing_if_record_is_not_literature():
-    schema = load_schema('hep')
-    subschema = schema['properties']['abstracts']
-
-    record = {
-        '$schema': 'http://localhost:5000/schemas/records/other.json',
-        'abstracts': [
-            {
-                'source': 'foo',
-                'value': 'bar',
-            },
-        ],
-    }
-    assert validate(record['abstracts'], subschema) is None
-
-    populate_abstract_source_suggest(None, record)
-
-    expected = [
-        {
-            'source': 'foo',
-            'value': 'bar',
-        },
-    ]
-    result = record['abstracts']
-
-    assert expected == result
-
-
-def test_populate_title_suggest_with_all_inputs():
-    schema = load_schema('journals')
-    journal_title_schema = schema['properties']['journal_title']
-    short_title_schema = schema['properties']['short_title']
-    title_variants_schema = schema['properties']['title_variants']
-
-    record = {
-        '$schema': 'http://localhost:5000/schemas/records/journals.json',
-        'journal_title': {'title': 'The Journal of High Energy Physics (JHEP)'},
-        'short_title': 'JHEP',
-        'title_variants': ['JOURNAL OF HIGH ENERGY PHYSICS'],
-    }
-    assert validate(record['journal_title'], journal_title_schema) is None
-    assert validate(record['short_title'], short_title_schema) is None
-    assert validate(record['title_variants'], title_variants_schema) is None
-
-    populate_title_suggest(None, record)
-
-    expected = {
-        'input': [
-            'The Journal of High Energy Physics (JHEP)',
-            'JHEP',
-            'JOURNAL OF HIGH ENERGY PHYSICS'
-        ],
-    }
-
-    result = record['title_suggest']
-
-    assert expected == result
-
-
-def test_populate_title_suggest_does_nothing_if_record_is_not_journal():
-    record = {'$schema': 'http://localhost:5000/schemas/records/other.json'}
-
-    populate_title_suggest(None, record)
-
-    assert 'title_suggest' not in record
-
-
-def test_populate_affiliation_suggest_from_icn():
-    schema = load_schema('institutions')
-    subschema = schema['properties']['ICN']
-
-    record = {
-        '$schema': 'http://localhost:5000/schemas/records/institutions.json',
-        'ICN': [
-            'CERN, Geneva',
-        ],
-        'legacy_ICN': 'CERN',
-    }
-    assert validate(record['ICN'], subschema) is None
-
-    populate_affiliation_suggest(None, record)
-
-    expected = {
-        'input': [
-            'CERN, Geneva',
-            'CERN',
-        ],
-    }
-    result = record['affiliation_suggest']
-
-    assert expected == result
-
-
-def test_populate_affiliation_suggest_from_institution_hierarchy_acronym():
-    schema = load_schema('institutions')
-    subschema = schema['properties']['institution_hierarchy']
-
-    record = {
-        '$schema': 'http://localhost:5000/schemas/records/institutions.json',
-        'institution_hierarchy': [
-            {'acronym': 'CERN'},
-        ],
-        'legacy_ICN': 'CERN',
-    }
-    assert validate(record['institution_hierarchy'], subschema) is None
-
-    populate_affiliation_suggest(None, record)
-
-    expected = {
-        'input': [
-            'CERN',
-            'CERN',
-        ],
-    }
-    result = record['affiliation_suggest']
-
-    assert expected == result
-
-
-def test_populate_affiliation_suggest_from_institution_hierarchy_name():
-    schema = load_schema('institutions')
-    subschema = schema['properties']['legacy_ICN']
-
-    record = {
-        '$schema': 'http://localhost:5000/schemas/records/institutions.json',
-        'institution_hierarchy': [
-            {'name': 'European Organization for Nuclear Research'},
-        ],
-        'legacy_ICN': 'CERN',
-    }
-    assert validate(record['legacy_ICN'], subschema) is None
-
-    populate_affiliation_suggest(None, record)
-
-    expected = {
-        'input': [
-            'European Organization for Nuclear Research',
-            'CERN',
-        ],
-    }
-    result = record['affiliation_suggest']
-
-    assert expected == result
-
-
-def test_populate_affiliation_suggest_from_legacy_icn():
-    schema = load_schema('institutions')
-    subschema = schema['properties']['legacy_ICN']
-
-    record = {
-        '$schema': 'http://localhost:5000/schemas/records/institutions.json',
-        'legacy_ICN': 'CERN',
-    }
-    assert validate(record['legacy_ICN'], subschema) is None
-
-    populate_affiliation_suggest(None, record)
-
-    expected = {
-        'input': [
-            'CERN',
-        ],
-    }
-    result = record['affiliation_suggest']
-
-    assert expected == result
-
-
-def test_populate_affiliation_suggest_from_name_variants():
-    schema = load_schema('institutions')
-    subschema = schema['properties']['name_variants']
-
-    record = {
-        '$schema': 'http://localhost:5000/schemas/records/institutions.json',
-        'legacy_ICN': 'CERN',
-        'name_variants': [
-            {'value': u'Centre Européen de Recherches Nucléaires'},
-        ],
-    }
-    assert validate(record['name_variants'], subschema) is None
-
-    populate_affiliation_suggest(None, record)
-
-    expected = {
-        'input': [
-            'CERN',
-            u'Centre Européen de Recherches Nucléaires',
-        ],
-    }
-    result = record['affiliation_suggest']
-
-    assert expected == result
-
-
-def test_populate_affiliation_suggest_from_postal_code():
-    schema = load_schema('institutions')
-    subschema = schema['properties']['addresses']
-
-    record = {
-        '$schema': 'http://localhost:5000/schemas/records/institutions.json',
-        'addresses': [
-            {'postal_code': '1211'},
-        ],
-        'legacy_ICN': 'CERN',
-    }
-    assert validate(record['addresses'], subschema) is None
-
-    populate_affiliation_suggest(None, record)
-
-    expected = {
-        'input': [
-            'CERN',
-            '1211',
-        ],
-    }
-    result = record['affiliation_suggest']
-
-    assert expected == result
-
-
-def test_populate_affiliation_suggest_to_ref():
-    schema = load_schema('institutions')
-    subschema = schema['properties']['self']
-
-    record = {
-        '$schema': 'http://localhost:5000/schemas/records/institutions.json',
-        'legacy_ICN': 'CERN',
-        'self': {
-            '$ref': 'http://localhost:5000/api/institutions/902725',
-        },
-    }
-    assert validate(record['self'], subschema) is None
-
-    populate_affiliation_suggest(None, record)
-
-    expected = {
-        'input': [
-            'CERN',
-        ],
-    }
-    result = record['affiliation_suggest']
-
-    assert expected == result
-
-
-def test_populate_affiliation_suggest_does_nothing_if_record_is_not_institution():
-    record = {'$schema': 'http://localhost:5000/schemas/records/other.json'}
-
-    populate_affiliation_suggest(None, record)
-
-    assert 'affiliation_suggest' not in record
-
-
-def test_populate_author_count():
-    schema = load_schema('hep')
-    subschema = schema['properties']['authors']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'authors': [
-            {
-                'full_name': 'Smith, John',
-                'inspire_roles': [
-                    'author',
-                ],
-            },
-            {
-                'full_name': 'Rafelski, Johann',
-                'inspire_roles': [
-                    'author',
-                    'editor',
-                ],
-            },
-            {
-                'full_name': 'Rohan, George',
-                'inspire_roles': [
-                    'author',
-                    'supervisor',
-                ],
-            },
-        ],
-    }
-    assert validate(record['authors'], subschema) is None
-
-    populate_author_count(None, record)
-
-    assert record['author_count'] == 2
-
-
-def test_populate_author_count_does_nothing_if_record_is_not_literature():
-    record = {'$schema': 'http://localhost:5000/schemas/records/other.json'}
-
-    populate_author_count(None, record)
-
-    assert 'author_count' not in record
-
-
-def test_populate_authors_full_name_unicode_normalized():
-    schema = load_schema('hep')
-    subschema = schema['properties']['authors']
-
-    record = {
-        '$schema': 'http://localhost:5000/records/schemas/hep.json',
-        'authors': [
-            {
-                'full_name': u'Müller, J.',
-            },
-            {
-                'full_name': u'Muller, J.',
-            },
-        ],
-    }
-    assert validate(record['authors'], subschema) is None
-
-    populate_authors_full_name_unicode_normalized(None, record)
-
-    expected = [
-        {
-            'full_name': u'Müller, J.',
-            'full_name_unicode_normalized': u'müller, j.',
-        },
-        {
-            'full_name': u'Muller, J.',
-            'full_name_unicode_normalized': u'muller, j.',
-        },
-    ]
-    result = record['authors']
-
-    assert expected == result
-
-
-def test_populate_authors_full_name_unicode_normalized_does_nothing_if_record_is_not_literature():
-    record = {
-        '$schema': 'http://localhost:5000/schemas/records/other.json',
-        'authors': [
-            {
-                'full_name': u'Müller, J.',
-            },
-            {
-                'full_name': u'Muller, J.',
-            },
-        ],
-    }
-
-    populate_authors_full_name_unicode_normalized(None, record)
-
-    expected = [
-        {
-            'full_name': u'Müller, J.',
-        },
-        {
-            'full_name': u'Muller, J.',
-        },
-    ]
-    result = record['authors']
-
     assert expected == result
